% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foodaccess}
\alias{foodaccess}
\title{USDA Food Access Research Atlas}
\format{
\subsection{\code{foodaccess}}{

A data frame with 72,531 rows and 147 columns
\describe{
\item{census_tract}{Census tract number}
\item{state}{State name}
\item{county}{County name}
\item{urban_flag}{Flag for urban tract}
\item{pop2010}{Population count from 2010 census}
\item{ohu2010}{Occupied housing unit count from 2010 census}
\item{group_quarters_flag}{Flag for tract where >=67\%}
\item{num_in_group_quarters}{Count of tract population residing in group quarters}
\item{pct_in_group_quarters}{Percent of tract population residing in group quarters}
\item{li_la_1_10}{Flag for low-income and low access when considering low accessibilty at 1 and 10 miles}
\item{li_la_half_10}{Flag for low-income and low access when considering low accessibilty at 1/2 and 10 miles}
\item{li_la_1_20}{Flag for low-income and low access when considering low accessibilty at 1 and 20 miles}
\item{li_la_vehicle}{Flag for low-income and low access when considering vehicle access or at 20 miles}
\item{la_lva_flag}{Flag for tract where >= 100 of households do not have a vehicle, and beyond 1/2 mile from supermarket}
\item{low_income_tracts}{Flag for low income tract}
\item{poverty_rate}{Share of the tract population living with income at or below the Federal poverty thresholds for family size}
\item{median_family_income}{Tract median family income}
\item{la_1_10}{Flag for low access tract at 1 mile for urban areas or 10 miles for rural areas}
\item{la_half_10}{Flag for low access tract at 1/2 mile for urban areas or 10 miles for rural areas}
\item{la_1_20}{Flag for low access tract at 1 mile for urban areas or 20 miles for rural areas}
\item{la_tracts_half}{Flag for low access tract when considering 1/2 mile distance}
\item{la_tracts_1}{Flag for low access tract when considering 1 mile distance}
\item{la_tracts_10}{Flag for low access tract when considering 10 mile distance}
\item{la_tracts_20}{Flag for low access tract when considering 20 mile distance}
\item{la_tracts_vehicle_20}{Flag for tract where >= 100 of households do not have a vehicle, and beyond 1/2 mile from supermarket; or >= 500 individuals are beyond 20 miles from supermarket ; or >= 33\% of individuals are beyond 20 miles from supermarket}
\item{LAPOP1_10}{Population count beyond 1 mile for urban areas or 10 miles for rural areas from supermarket}
\item{LAPOP05_10}{Population count beyond 1/2 mile for urban areas or 10 miles for rural areas from supermarket}
\item{LAPOP1_20}{Population count beyond 1 mile for urban areas or 20 miles for rural areas from supermarket}
\item{LALOWI1_10}{Low income population count beyond 1 mile for urban areas or 10 miles for rural areas from supermarket}
\item{LALOWI05_10}{Low income population count beyond 1/2 mile for urban areas or 10 miles for rural areas from supermarket}
\item{LALOWI1_20}{Low income population count beyond 1 mile for urban areas or 20 miles for rural areas from supermarket}
\item{lapophalf}{Population count beyond 1/2 mile from supermarket}
\item{lapophalfshare}{Share of tract population that are beyond 1/2 mile from supermarket}
\item{lalowihalf}{Low income population count beyond 1/2 mile from supermarket}
\item{lalowihalfshare}{Share of tract population that are low income individuals beyond 1/2 mile from supermarket}
\item{lakidshalf}{Kids population count beyond 1/2 mile from supermarket}
\item{lakidshalfshare}{Share of tract population that are kids beyond 1/2 mile from supermarket}
\item{laseniorshalf}{Seniors population count beyond 1/2 mile from supermarket}
\item{laseniorshalfshare}{Share of tract population that are seniors beyond 1/2 mile from supermarket}
\item{lawhitehalf}{White population count beyond 1/2 mile from supermarket}
\item{lawhitehalfshare}{Share of tract population that are white beyond 1/2 mile from supermarket}
\item{lablackhalf}{Black or African American population count beyond 1/2 mile from supermarket}
\item{lablackhalfshare}{Share of tract population that are Black or African American beyond 1/2 mile from supermarket}
\item{laasianhalf}{Asian population count beyond 1/2 mile from supermarket}
\item{laasianhalfshare}{Share of tract population that are Asian beyond 1/2 mile from supermarket}
\item{lanhopihalf}{Native Hawaiian or Other Pacific Islander population count beyond 1/2 mile from supermarket}
\item{lanhopihalfshare}{Share of tract population that are Native Hawaiian or Other Pacific Islander beyond 1/2 mile from supermarket}
\item{laaianhalf}{American Indian or Alaska Native population count beyond 1/2 mile from supermarket}
\item{laaianhalfshare}{Share of tract population that are American Indian or Alaska Native beyond 1/2 mile from supermarket}
\item{laomultirhalf}{Other/Multiple race population count beyond 1/2 mile from supermarket}
\item{laomultirhalfshare}{Share of tract population that are Other/Multiple race beyond 1/2 mile from supermarket}
\item{lahisphalf}{Hispanic or Latino ethnicity population count beyond 1/2 mile from supermarket}
\item{lahisphalfshare}{Share of tract population that are of Hispanic or Latino ethnicity beyond 1/2 mile from supermarket}
\item{lahunvhalf}{Housing units without vehicle count beyond 1/2 mile from supermarket}
\item{lahunvhalfshare}{Share of tract housing units that are without vehicle and beyond 1/2 mile from supermarket}
\item{lasnaphalf}{Housing units receiving SNAP benefits count beyond 1/2 mile from supermarket}
\item{lasnaphalfshare}{Share of tract housing units receiving SNAP benefits count beyond 1/2 mile from supermarket}
\item{lapop1}{Population count beyond 1 mile from supermarket}
\item{lapop1share}{Share of tract population that are beyond 1 mile from supermarket}
\item{lalowi1}{Low income population count beyond 1 mile from supermarket}
\item{lalowi1share}{Share of tract population that are low income individuals beyond 1 mile from supermarket}
\item{lakids1}{Kids population count beyond 1 mile from supermarket}
\item{lakids1share}{Share of tract population that are kids beyond 1 mile from supermarket}
\item{laseniors1}{Seniors population count beyond 1 mile from supermarket}
\item{laseniors1share}{Share of tract population that are seniors beyond 1 mile from supermarket}
\item{lawhite1}{White population count beyond 1 mile from supermarket}
\item{lawhite1share}{Share of tract population that are white beyond 1 mile from supermarket}
\item{lablack1}{Black or African American population count beyond 1 mile from supermarket}
\item{lablack1share}{Share of tract population that are Black or African American beyond 1 mile from supermarket}
\item{laasian1}{Asian population count beyond 1 mile from supermarket}
\item{laasian1share}{Share of tract population that are Asian beyond 1 mile from supermarket}
\item{lanhopi1}{Native Hawaiian or Other Pacific Islander population count beyond 1 mile from supermarket}
\item{lanhopi1share}{Share of tract population that are Native Hawaiian or Other Pacific Islander beyond 1 mile from supermarket}
\item{laaian1}{American Indian or Alaska Native population count beyond 1 mile from supermarket}
\item{laaian1share}{Share of tract population that are American Indian or Alaska Native beyond 1 mile from supermarket}
\item{laomultir1}{Other/Multiple race population count beyond 1 mile from supermarket}
\item{laomultir1share}{Share of tract population that are Other/Multiple race beyond 1 mile from supermarket}
\item{lahisp1}{Hispanic or Latino ethnicity population count beyond 1 mile from supermarket}
\item{lahisp1share}{Share of tract population that are of Hispanic or Latino ethnicity beyond 1 mile from supermarket}
\item{lahunv1}{Housing units without vehicle count beyond 1 mile from supermarket}
\item{lahunv1share}{Share of tract housing units that are without vehicle and beyond 1 mile from supermarket}
\item{lasnap1}{Housing units receiving SNAP benefits count beyond 1 mile from supermarket}
\item{lasnap1share}{Share of tract housing units receiving SNAP benefits count beyond 1 mile from supermarket}
\item{lapop10}{Population count beyond 10 miles from supermarket}
\item{lapop10share}{Share of tract population that are beyond 10 miles from supermarket}
\item{lalowi10}{Low income population count beyond 10 miles from supermarket}
\item{lalowi10share}{Share of tract population that are low income individuals beyond 10 miles from supermarket}
\item{lakids10}{Kids population count beyond 10 miles from supermarket}
\item{lakids10share}{Share of tract population that are kids beyond 10 miles from supermarket}
\item{laseniors10}{Seniors population count beyond 10 miles from supermarket}
\item{laseniors10share}{Share of tract population that are seniors beyond 10 miles from supermarket}
\item{lawhite10}{White population count beyond 10 miles from supermarket}
\item{lawhite10share}{Share of tract population that are white beyond 10 miles from supermarket}
\item{lablack10}{Black or African American population count beyond 10 miles from supermarket}
\item{lablack10share}{Share of tract population that are Black or African American beyond 10 miles from supermarket}
\item{laasian10}{Asian population count beyond 10 miles from supermarket}
\item{laasian10share}{Share of tract population that are Asian beyond 10 miles from supermarket}
\item{lanhopi10}{Native Hawaiian or Other Pacific Islander population count beyond 10 miles from supermarket}
\item{lanhopi10share}{Share of tract population that are Native Hawaiian or Other Pacific Islander beyond 10 miles from supermarket}
\item{laaian10}{American Indian or Alaska Native population count beyond 10 miles from supermarket}
\item{laaian10share}{Share of tract population that are American Indian or Alaska Native beyond 10 miles from supermarket}
\item{laomultir10}{Other/Multiple race population count beyond 10 miles from supermarket}
\item{laomultir10share}{Share of tract population that are Other/Multiple race beyond 10 miles from supermarket}
\item{lahisp10}{Hispanic or Latino ethnicity population count beyond 10 miles from supermarket}
\item{lahisp10share}{Share of tract population that are of Hispanic or Latino ethnicity beyond 10 miles from supermarket}
\item{lahunv10}{Housing units without vehicle count beyond 10 miles from supermarket}
\item{lahunv10share}{Share of tract housing units that are without vehicle and beyond 10 miles from supermarket}
\item{lasnap10}{Housing units receiving SNAP benefits count beyond 10 miles from supermarket}
\item{lasnap10share}{Share of tract housing units receiving SNAP benefits count beyond 10 miles from supermarket}
\item{lapop20}{Population count beyond 20 miles from supermarket}
\item{lapop20share}{Share of tract population that are beyond 20 miles from supermarket}
\item{lalowi20}{Low income population count beyond 20 miles from supermarket}
\item{lalowi20share}{Share of tract population that are low income individuals beyond 20 miles from supermarket}
\item{lakids20}{Kids population count beyond 20 miles from supermarket}
\item{lakids20share}{Share of tract population that are kids beyond 20 miles from supermarket}
\item{laseniors20}{Seniors population count beyond 20 miles from supermarket}
\item{laseniors20share}{Share of tract population that are seniors beyond 20 miles from supermarket}
\item{lawhite20}{White population count beyond 20 miles from supermarket}
\item{lawhite20share}{Share of tract population that are white beyond 20 miles from supermarket}
\item{lablack20}{Black or African American population count beyond 20 miles from supermarket}
\item{lablack20share}{Share of tract population that are Black or African American beyond 20 miles from supermarket}
\item{laasian20}{Asian population count beyond 20 miles from supermarket}
\item{laasian20share}{Share of tract population that are Asian beyond 20 miles from supermarket}
\item{lanhopi20}{Native Hawaiian or Other Pacific Islander population count beyond 20 miles from supermarket}
\item{lanhopi20share}{Share of tract population that are Native Hawaiian or Other Pacific Islander beyond 20 miles from supermarket}
\item{laaian20}{American Indian or Alaska Native population count beyond 20 miles from supermarket}
\item{laaian20share}{Share of tract population that are American Indian or Alaska Native beyond 20 miles from supermarket}
\item{laomultir20}{Other/Multiple race population count beyond 20 miles from supermarket}
\item{laomultir20share}{Share of tract population that are Other/Multiple race beyond 20 miles from supermarket}
\item{lahisp20}{Hispanic or Latino ethnicity population count beyond 20 miles from supermarket}
\item{lahisp20share}{Share of tract population that are of Hispanic or Latino ethnicity beyond 20 miles from supermarket}
\item{lahunv20}{Housing units without vehicle count beyond 20 miles from supermarket}
\item{lahunv20share}{Share of tract housing units that are without vehicle and beyond 20 miles from supermarket}
\item{lasnap20}{Housing units receiving SNAP benefits count beyond 20 miles from supermarket}
\item{lasnap20share}{Share of tract housing units receiving SNAP benefits count beyond 20 miles from supermarket}
\item{TractLOWI}{Total count of low-income population in tract}
\item{TractKids}{Total count of children age 0-17 in tract}
\item{TractSeniors}{Total count of seniors age 65+ in tract}
\item{TractWhite}{Total count of White population in tract}
\item{TractBlack}{Total count of Black or African American population in tract}
\item{TractAsian}{Total count of Asian population in tract}
\item{TractNHOPI}{Total count of Native Hawaiian and Other Pacific Islander population in tract}
\item{TractAIAN}{Total count of American Indian and Alaska Native population in tract}
\item{TractOMultir}{Total count of Other/Multiple race population in tract}
\item{TractHispanic}{Total count of Hispanic or Latino population in tract}
\item{TractHUNV}{Total count of housing units without a vehicle in tract}
\item{TractSNAP}{Total count of housing units receiving SNAP benefits in tract}
}
}
}
\source{
\url{https://www.ers.usda.gov/data-products/food-access-research-atlas/download-the-data}
}
\usage{
foodaccess
}
\description{
A subset of data from the USDA's Food Access Research Atlas dataset
}
\keyword{datasets}
